+++ packages/cldr-core/__tests__/systems/calendars/weeks.test.ts
@@ -25,40 +25,86 @@ test('weeks coverage', () => {
   expect(d.weekOfMonth()).toEqual(5);
   expect(d.weekOfYear()).toEqual(1);
   expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(53);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
 
   // Dec 30
   d = d.subtract({ day: 1 });
   expect(d.weekOfMonth()).toEqual(5);
   expect(d.weekOfYear()).toEqual(1);
   expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(53);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
 
   // Dec 29
   d = d.subtract({ day: 1 });
   expect(d.weekOfMonth()).toEqual(5);
   expect(d.weekOfYear()).toEqual(1);
   expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(53);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
 
   // Dec 28
   d = d.subtract({ day: 1 });
   expect(d.weekOfMonth()).toEqual(5);
   expect(d.weekOfYear()).toEqual(1);
   expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(53);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
 
   // Dec 27
   d = d.subtract({ day: 1 });
   expect(d.weekOfMonth()).toEqual(5);
   expect(d.weekOfYear()).toEqual(1);
   expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(52);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
 
   // Dec 26
   d = d.subtract({ day: 1 });
   expect(d.weekOfMonth()).toEqual(4);
   expect(d.weekOfYear()).toEqual(52);
   expect(d.yearOfWeekOfYear()).toEqual(2020);
+  expect(d.weekOfYearISO()).toEqual(52);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
 
   // Dec 25
   d = d.subtract({ day: 1 });
   expect(d.weekOfMonth()).toEqual(4);
   expect(d.weekOfYear()).toEqual(52);
   expect(d.yearOfWeekOfYear()).toEqual(2020);
+  expect(d.weekOfYearISO()).toEqual(52);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
+
+  // Jan 1
+  d = make(n, NEW_YORK).add({ day: 1 });
+  expect(d.weekOfMonth()).toEqual(1);
+  expect(d.weekOfYear()).toEqual(1);
+  expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(53);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
+
+  // Jan 2
+  d = d.add({ day: 1 });
+  expect(d.weekOfMonth()).toEqual(1);
+  expect(d.weekOfYear()).toEqual(1);
+  expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(53);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
+
+  // Jan 3
+  d = d.add({ day: 1 });
+  expect(d.weekOfMonth()).toEqual(2);
+  expect(d.weekOfYear()).toEqual(2);
+  expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(53);
+  expect(d.yearOfWeekofYearISO()).toEqual(2020);
+
+  // Jan 4
+  d = d.add({ day: 1 });
+  expect(d.weekOfMonth()).toEqual(2);
+  expect(d.weekOfYear()).toEqual(2);
+  expect(d.yearOfWeekOfYear()).toEqual(2021);
+  expect(d.weekOfYearISO()).toEqual(1);
+  expect(d.yearOfWeekofYearISO()).toEqual(2021);
 });